requirements.txt
PyQt6>=6.4.0
pyserial>=3.5
pywin32>=306; sys_platform == "win32"
pyinstaller>=6.0.0
